Rearrange types in different files
