+++ pages/api/data-explorer/species-planted.ts
@@ -7,6 +7,7 @@ import {
 } from '../../../src/middlewares/rate-limiter';
 import NodeCache from 'node-cache';
 import { getCachedKey } from '../../../src/utils/getCachedKey';
+import { ISpeciesPlanted } from '../../../src/features/common/types/dataExplorer';
 
 const ONE_HOUR_IN_SEC = 60 * 60;
 const ONE_DAY = ONE_HOUR_IN_SEC * 24;
@@ -17,13 +18,6 @@ const handler = nc<NextApiRequest, NextApiResponse>();
 
 handler.use(rateLimiter);
 handler.use(speedLimiter);
-
-export interface ISpeciesPlanted {
-  other_species: null | string;
-  scientific_species_id: number;
-  name: null | string;
-  total_tree_count: number;
-}
 
 handler.post(async (req, response) => {
   const { projectId, startDate, endDate } = req.body;
