+++ src/features/common/types/dataExplorer.d.ts
@@ -22,3 +22,35 @@ export interface IYearlyFrame {
   year: number;
   treesPlanted: number;
 }
+
+export interface ISpeciesPlanted {
+  other_species: null | string;
+  scientific_species_id: number;
+  name: null | string;
+  total_tree_count: number;
+}
+
+export interface IExportData {
+  hid: string;
+  plant_date: Date;
+  species: string;
+  tree_count: number;
+  geometry: string;
+  type: string;
+  trees_allocated: number;
+  trees_planted: number;
+  metadata: string;
+  description: null;
+  plant_project_id: number;
+  sample_tree_count: number;
+  capture_status: string;
+  created: Date;
+}
+
+export interface TotalTreesPlanted {
+  totalTreesPlanted: number;
+}
+
+export interface TotalSpeciesPlanted {
+  totalSpeciesPlanted: number;
+}
